requirements.txt
numpy
paddlenlp
paddlepaddle-gpu
scikit-learn
tqdm
matplotlib
IPython
pre-commit